change temporal wrapper to util
