+++ terratorch/models/utils.py
@@ -1,7 +1,7 @@
-import logging
-
 from torch import nn, Tensor
 import torch 
+from terratorch.registry import BACKBONE_REGISTRY
+import pdb
 
 class DecoderNotFoundError(Exception):
     pass
@@ -53,3 +53,122 @@ def pad_images(imgs: Tensor, patch_size: int | list, padding: str) -> Tensor:
             for img in imgs  # Apply per image to avoid NotImplementedError from torch.nn.functional.pad
         ])
     return imgs
+
+
+def _get_backbone(backbone: str | nn.Module, **backbone_kwargs) -> nn.Module:
+    use_temporal = backbone_kwargs.pop('use_temporal', None)
+    temporal_pooling = backbone_kwargs.pop('temporal_pooling', None)
+    concat = backbone_kwargs.pop('temporal_concat', None)
+    if isinstance(backbone, nn.Module):
+        model = backbone
+    else:
+        model = BACKBONE_REGISTRY.build(backbone, **backbone_kwargs)
+
+    # Apply TemporalWrapper inside _get_backbone
+    if use_temporal:
+        model = TemporalWrapper(model, pooling=temporal_pooling, concat=concat)
+
+    return model
+
+def subtract_along_dim2(tensor: torch.Tensor):
+    return tensor[:, :, 0, ...] - tensor[:, :, 1, ...]
+
+
+class TemporalWrapper(nn.Module):
+    def __init__(self, encoder: nn.Module, pooling="mean", concat=False, n_timestamps=None):
+        """
+        Wrapper for applying a temporal encoder across multiple time steps.
+
+        Args:
+            encoder (nn.Module): The feature extractor (backbone).
+            pooling (str): Type of pooling ('mean', 'max', 'diff') with 'diff' working only with 2 timestamps.
+            concat (bool): Whether to concatenate features instead of pooling.
+            n_timestamps (int): Number of timestamps. Necessary in case of concat.
+        """
+        super().__init__()
+        self.encoder = encoder
+        self.concat = concat
+        self.pooling_type = pooling
+        self.n_timestamps = n_timestamps
+
+        if pooling not in ["mean", "max", "diff"]:
+            raise ValueError("Pooling must be 'mean', 'max' or 'diff'")
+
+        # Ensure the encoder has an out_channels attribute
+        if hasattr(encoder, "out_channels"):
+            self.out_channels = encoder.out_channels * (1 if not concat else self.n_timestamps)
+        else:
+            raise AttributeError("Encoder must have an `out_channels` attribute.")
+
+    def forward(self, x: torch.Tensor) -> list[torch.Tensor]:
+        """
+        Forward pass for temporal processing.
+
+        Args:
+            x (Tensor): Input tensor of shape [B, C, T, H, W].
+
+        Returns:
+            List[Tensor]: A list of processed tensors, one per feature map.
+        """
+        
+        if x.dim() != 5:
+            raise ValueError(f"Expected input shape [B, C, T, H, W], but got {x.shape}")
+            
+        if (self.pooling_type == 'diff') & (x.shape[2] != 2):
+            raise ValueError(f"Expected 2 timestamps for aggregation method 'diff'")
+
+        batch_size, channels, timesteps, height, width = x.shape
+
+        # Initialize lists to store feature maps at each timestamp
+        num_feature_maps = None  # Will determine dynamically
+        features_per_map = []  # Stores feature maps across timestamps
+        
+        for t in range(timesteps):
+            feat = self.encoder(x[:, :, t, :, :])  # Extract features at timestamp t
+
+            if not isinstance(feat, list):  # If the encoder outputs a single feature map, convert to list
+                if isinstance(feat, tuple):
+                    feat = list(feat)
+                else:
+                    feat = [feat]
+
+            if num_feature_maps is None:
+                num_feature_maps = len(feat)  # Determine how many feature maps the encoder produces
+
+            for i, feature_map in enumerate(feat):
+                if len(features_per_map) <= i:
+                    features_per_map.append([])  # Create list for each feature map
+
+                feature_map = feature_map[0] if isinstance(feature_map, tuple) else feature_map
+                features_per_map[i].append(feature_map)  # Store feature map at time t
+
+        # Stack features along the temporal dimension
+        for i in range(num_feature_maps):
+            try:
+                features_per_map[i] = torch.stack(features_per_map[i], dim=2)  # Shape: [B, C', T, H', W']
+            except RuntimeError as e:
+                raise
+
+        # Apply pooling or concatenation
+        if self.concat:
+            features_per_map_agg = [feat.view(batch_size, -1, feat.shape[-2], feat.shape[-1]) if len(feat.shape) == 5 else feat.view(batch_size, feat.shape[-3], -1) for feat in features_per_map]
+        elif self.pooling_type == "max":
+            features_per_map_agg = [torch.max(feat, dim=2)[0] for feat in features_per_map]  # Max pooling across T
+        elif self.pooling_type == "diff":
+            features_per_map_agg = [feat[:, :, 0, ...] - feat[:, :, 1, ...] for feat in features_per_map]
+        else:
+            features_per_map_agg = [torch.mean(feat, dim=2) for feat in features_per_map]
+        return features_per_map_agg
+#         print("original ", features_per_map[0].shape)
+#         features_per_map_agg = [feat.view(batch_size, -1, feat.shape[-2], feat.shape[-1]) if len(feat.shape) == 5 else feat.view(batch_size, feat.shape[-3], -1) for feat in features_per_map]
+#         print("concat ", features_per_map_agg[0].shape)
+#         features_per_map_agg = [torch.max(feat, dim=2)[0] for feat in features_per_map]  # Max pooling across T
+#         print("max ", features_per_map_agg[0].shape)
+#         features_per_map_agg = [feat[:, :, 0, ...] - feat[:, :, 1, ...] for feat in features_per_map]
+#         print("diff ", features_per_map_agg[0].shape)
+#         features_per_map_agg = [torch.mean(feat, dim=2) for feat in features_per_map]
+#         print("mean ", features_per_map_agg[0].shape)
+
+#         pdb.set_trace()
+#         return features_per_map_agg
+
